solve authMiddleware errors

+++ MenuCategory/views.py
@@ -1,53 +1,52 @@
-from rest_framework import status
+from rest_framework.decorators import api_view, authentication_classes, permission_classes
 from rest_framework.response import Response
-from rest_framework.views import APIView
+from rest_framework import status
+from django.shortcuts import get_object_or_404
 from .models import MenuCategory
 from .serializers import MenuCategorySerializer
+from rest_framework.authentication import SessionAuthentication, TokenAuthentication
+from rest_framework.permissions import IsAuthenticated
+from users.middleware import AdminRoleMiddleware
 
-class MenuCategoryList(APIView):
-    """
-    List all menu categories, or create a new menu category.
-    """
-    def get(self, request):
+@api_view(['GET', 'POST'])
+@authentication_classes([SessionAuthentication, TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def menu_category_list(request):
+    admin_middleware = AdminRoleMiddleware(None)  # Apply AdminRoleMiddleware
+    response = admin_middleware(request)
+    if response:
+        return response
+
+    if request.method == 'GET':
         menu_categories = MenuCategory.objects.all()
         serializer = MenuCategorySerializer(menu_categories, many=True)
         return Response(serializer.data)
-    
-    def post(self, request):
+    elif request.method == 'POST':
         serializer = MenuCategorySerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class MenuCategoryDetail(APIView):
-    """
-    Retrieve, update or delete a menu category instance.
-    """
-    def get_object(self, pk):
-        try:
-            return MenuCategory.objects.get(pk=pk)
-        except MenuCategory.DoesNotExist:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
-            #raise Http404
-    
-    def get(self, request, pk):
-        menu_category = self.get_object(pk)
+@api_view(['GET', 'PUT', 'DELETE'])
+@authentication_classes([SessionAuthentication, TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def menu_category_detail(request, pk):
+    admin_middleware = AdminRoleMiddleware(None)  # Apply AdminRoleMiddleware
+    response = admin_middleware(request)
+    if response:
+        return response
+
+    menu_category = get_object_or_404(MenuCategory, pk=pk)
+    if request.method == 'GET':
         serializer = MenuCategorySerializer(menu_category)
         return Response(serializer.data)
-    
-    def put(self, request, pk):
-        menu_category = self.get_object(pk)
+    elif request.method == 'PUT':
         serializer = MenuCategorySerializer(menu_category, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-    
-    def delete(self, request, pk):
-        menu_category = self.get_object(pk)
+    elif request.method == 'DELETE':
         menu_category.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
-    
-
-
