+++ users/views.py
@@ -1,50 +1,59 @@
-from rest_framework import status
+from rest_framework.decorators import api_view, authentication_classes, permission_classes
 from rest_framework.response import Response
-from rest_framework.views import APIView
+from rest_framework import status
+from django.utils.decorators import method_decorator
+from django.shortcuts import get_object_or_404
 from .models import User
 from .serializers import UserSerializer
+from rest_framework.authentication import SessionAuthentication, TokenAuthentication
+from rest_framework.permissions import IsAuthenticated
+from users.middleware import AdminRoleMiddleware
 
-class UserList(APIView):
-    """
-    List all users or create a new user.
-    """
-    def get(self, request):
+@api_view(['GET', 'POST'])
+@authentication_classes([SessionAuthentication, TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def user_list(request):
+    admin_middleware = AdminRoleMiddleware(None)  # Initialize with None as get_response
+    response = admin_middleware(request)
+    if response:
+        return response
+    if request.method == 'GET':
         users = User.objects.all()
         serializer = UserSerializer(users, many=True)
         return Response(serializer.data)
-
-    def post(self, request):
+    
+    elif request.method == 'POST':
         serializer = UserSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_201_CREATED)
+            username = request.data.get('username')  # Retrieve username from request data
+            password = request.data.get('password')  # Retrieve password from request data
+            email = request.data.get('email')
+            role = request.data.get('role')
+            if not username or not password:
+                return Response({'error': 'Username and password are required'}, status=status.HTTP_400_BAD_REQUEST)
+            
+            # Create a new user with the provided username and password
+            user = User.objects.create_user(username=username, password=password, email=email, role=role)
+            return Response({"user": serializer.data}, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class UserDetail(APIView):
-    """
-    Retrieve, update or delete a user instance.
-    """
-    def get_object(self, pk):
-        try:
-            return User.objects.get(pk=pk)
-        except User.DoesNotExist:
-            return Response(status=status.HTTP_400_BAD_REQUEST)
-            #raise Http404
-
-    def get(self, request, pk):
-        user = self.get_object(pk)
+@api_view(['GET', 'PUT', 'DELETE'])
+@authentication_classes([SessionAuthentication, TokenAuthentication])
+@permission_classes([IsAuthenticated])
+def user_detail(request, pk):
+    user = get_object_or_404(User, pk=pk)
+    
+    if request.method == 'GET':
         serializer = UserSerializer(user)
         return Response(serializer.data)
-
-    def put(self, request, pk):
-        user = self.get_object(pk)
+    
+    elif request.method == 'PUT':
         serializer = UserSerializer(user, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-    def delete(self, request, pk):
-        user = self.get_object(pk)
+    
+    elif request.method == 'DELETE':
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
